refactor: enhance DDD structure
