+++ lib/features/auth/presentations/register_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:planthis/app/theme/app_colors.dart';
-import 'package:planthis/features/auth/presentation/widgets/password_requirements.dart';
-import '../logic/auth_controller.dart';
+import 'package:planthis/app/utils.dart';
+import 'package:planthis/features/auth/presentations/widgets/password_requirements.dart';
+import 'package:planthis/features/auth/providers/auth_provider.dart';
 
 class RegisterScreen extends ConsumerStatefulWidget {
   const RegisterScreen({super.key});
@@ -16,24 +17,11 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
   final passwordController = TextEditingController();
   final confirmPasswordController = TextEditingController();
 
-  String? errorMessage;
-
-  bool hasMinLength(String password) => password.length >= 8;
-  bool hasNumber(String password) => password.contains(RegExp(r'[0-9]'));
-  bool passwordsMatch(String password, String confirmPassword) =>
-      password == confirmPassword && password.isNotEmpty;
-
-  bool isValidEmail(String email) =>
-      RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$').hasMatch(email);
-
-  bool isPasswordValid(String password, String confirmPassword) {
-    return hasMinLength(password) &&
-        hasNumber(password) &&
-        passwordsMatch(password, confirmPassword);
-  }
-
   @override
   Widget build(BuildContext context) {
+    final authState = ref.watch(authControllerProvider);
+    final authController = ref.read(authControllerProvider.notifier);
+
     return Scaffold(
       appBar: AppBar(title: const Text('Register')),
       body: Padding(
@@ -60,22 +48,19 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
             const SizedBox(height: 4),
             PasswordRequirement(
               requirement: 'At least 8 characters',
-              isValid: hasMinLength(passwordController.text),
+              isValid: (passwordController.text).hasMinLength(8),
             ),
             PasswordRequirement(
               requirement: 'Contains a number',
-              isValid: hasNumber(passwordController.text),
+              isValid: (passwordController.text).hasNumber(),
             ),
             PasswordRequirement(
               requirement: 'Passwords match',
-              isValid: passwordsMatch(
-                passwordController.text,
-                confirmPasswordController.text,
-              ),
+              isValid: AppUtils.passwordMatch(passwordController.text, confirmPasswordController.text)
             ),
-            if (errorMessage != null) ...[
+            if (authState.hasError) ...[
               const SizedBox(height: 4),
-              Text(errorMessage!, style: const TextStyle(color: Colors.red)),
+              Text(authState.error.toString(), style: const TextStyle(color: Colors.red)),
             ],
             const SizedBox(height: 20),
             ElevatedButton(
@@ -84,25 +69,7 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
                 final password = passwordController.text.trim();
                 final confirmPassword = confirmPasswordController.text.trim();
 
-                if (email.isEmpty ||
-                    password.isEmpty ||
-                    confirmPassword.isEmpty) {
-                  setState(() => errorMessage = 'All fields are required.');
-                  return;
-                }
-                if (!isValidEmail(email)) {
-                  setState(() => errorMessage = 'Enter a valid email address.');
-                  return;
-                }
-                if (!isPasswordValid(password, confirmPassword)) {
-                  setState(() => errorMessage = 'Password rules not met.');
-                  return;
-                }
-
-                setState(() => errorMessage = null);
-                ref
-                    .read(authControllerProvider.notifier)
-                    .register(email, password);
+                authController.register(email, password, confirmPassword);
               },
               child: const Text('Register'),
             ),
