+++ lib/features/auth/presentations/widgets/password_requirements.dart
